+++ react/components/Table/index.js
@@ -26,8 +26,14 @@ class Table extends PureComponent {
       tableRowHeight: this.getRowHeight(props.density),
       selectedDensity: props.density,
       allChecked: false,
-      selectedRows: [],
+      selectedRows: this.props.bulkActions.selectedRows || [],
       allLinesSelected: false,
+    }
+  }
+
+  componentDidUpdate(prevProps) {
+    if(this.props.bulkActions && this.props.bulkActions.selectedRows) {
+      this.setState({selectedRows: this.props.bulkActions.selectedRows})
     }
   }
 
@@ -450,6 +456,7 @@ Table.propTypes = {
     }),
     totalItems: PropTypes.number,
     onChange: PropTypes.func,
+    selectedRows: PropTypes.array,
     main: PropTypes.shape({
       label: PropTypes.string.isRequired,
       handleCallback: PropTypes.func.isRequired,