+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Optional prop `selectedRows` that can control the table selected items
+
 ## [8.72.7] - 2019-08-13
 
 ### Changed